#5: refactoring
